add search functionality
