+++ Web/web_assessment/pages/doctor/[id].tsx
@@ -21,9 +21,9 @@ const DoctorDetailPage: React.FC = () => {
 
   return (
     <div>
-      <h1 className="text-2xl font-bold">{doctorProfile.name}</h1>
+      <h1 className="text-2xl font-bold">{doctorProfile.fullName}</h1>
       <div className="flex items-center mt-4">
-        <img src={doctorProfile.profilePicture} alt={doctorProfile.name} className="w-24 h-24 rounded-full" />
+        <img src={doctorProfile.photo} alt={doctorProfile.name} className="w-24 h-24 rounded-full" />
         <div className="ml-4">
           <p className="text-gray-500">Specialty: {doctorProfile.specialty}</p>
           <p className="text-gray-500">Email: {doctorProfile.email}</p>
